Fix: Fix type mismatch issue on terminal condition

+++ batchspec/backends/eagle_engine.py
@@ -1,7 +1,6 @@
 """EAGLE backend engine for speculative decoding with chain drafting."""
 
 from pathlib import Path
-from typing import Optional
 
 import torch
 from torch import Tensor
@@ -438,7 +437,7 @@ class EAGLEChainEngine(BaseEngine):
                 if (not force_budget) and eos_accepted_or_generated.any():
                     # On the terminal step, we need to write the bonus tokens to the output
                     terminal = True
-                    output[batch_indices, num_generated_tokens] = bonus_tokens[:, 0]
+                    output[batch_indices, num_generated_tokens] = bonus_tokens[:, 0].long()
                     num_generated_tokens += 1
         
         profiler.end_run()
